[Core] Fix facades being added to wrong multi-target framework
If .NET Framework is the first target framework listed then facade
assemblies would be added to all the other target frameworks. This
would cause text editor errors when .NET Standard was selected due to
extra facade assemblies being added to the Roslyn project. Now the
TargetFramework associated with the DotNetProjectConfiguration is
used when resolving facade assemblies.