[Core] WIP and try out a new interface that would just plug into FileWatcherService
V2 contains an API interface that allows batching subscriptions

It also tries to take into account that the file notifications we're looking at
are either:
a) Single file related - i.e. Project.Files
b) Glob based, so we'll be watching a directory recursively based on globs

This way, we can actually take advantages of a few things for pre-filtering:
a) Define directory watching by recursion based on globbing format
b) Define file watched based on the file tree

That way, we can deterministically reduce the number of cones being watched
and ensure we're only watching what we should.